+++ src/engine/input/controller.cpp
@@ -78,7 +78,7 @@ ControllerScroller*
 Controller::get_scroller(int id)
 {
   if (id >= 0 && id < int(scrollers.size()))
-    return scrollers[id].get();
+    return scrollers[static_cast<size_t>(id)].get();
   else
     return nullptr;
 }
@@ -87,17 +87,17 @@ void
 Controller::add_scroller(int id, std::unique_ptr<ControllerScroller> scroller)
 {
   if (int(scrollers.size())-1 < id)
-    scrollers.resize(id+1);
+    scrollers.resize(static_cast<size_t>(id+1));
 
-  assert(scrollers[id] == 0);
-  scrollers[id] = std::move(scroller);
+  assert(scrollers[static_cast<size_t>(id)] == 0);
+  scrollers[static_cast<size_t>(id)] = std::move(scroller);
 }
 
 ControllerPointer*
 Controller::get_pointer(int id)
 {
   if (id >= 0 && id < int(pointers.size()))
-    return pointers[id].get();
+    return pointers[static_cast<size_t>(id)].get();
   else
     return nullptr;
 }
@@ -106,7 +106,7 @@ ControllerKeyboard*
 Controller::get_keyboard(int id)
 {
   if (id >= 0 && id < int(keyboards.size()))
-    return keyboards[id].get();
+    return keyboards[static_cast<size_t>(id)].get();
   else
     return nullptr;
 }
@@ -115,44 +115,44 @@ void
 Controller::add_keyboard(int id, std::unique_ptr<ControllerKeyboard> keyboard)
 {
   if (int(keyboards.size())-1 < id)
-    keyboards.resize(id+1);
+    keyboards.resize(static_cast<size_t>(id+1));
 
-  assert(keyboards[id] == 0);
-  keyboards[id] = std::move(keyboard);
+  assert(keyboards[static_cast<size_t>(id)] == 0);
+  keyboards[static_cast<size_t>(id)] = std::move(keyboard);
 }
 
 void
 Controller::add_pointer(int id, std::unique_ptr<ControllerPointer> pointer)
 {
   if (int(pointers.size())-1 < id)
-    pointers.resize(id+1);
+    pointers.resize(static_cast<size_t>(id+1));
 
-  assert(pointers[id] == 0);
-  pointers[id] = std::move(pointer);
+  assert(pointers[static_cast<size_t>(id)] == 0);
+  pointers[static_cast<size_t>(id)] = std::move(pointer);
 }
 
 ControllerAxis*
 Controller::get_axis(int id)
 {
   assert(id >= 0 && id < int(axes.size()));
-  return axes[id].get();
+  return axes[static_cast<size_t>(id)].get();
 }
 
 void
 Controller::add_axis(int id, std::unique_ptr<ControllerAxis> axis)
 {
   if (int(axes.size())-1 < id)
-    axes.resize(id+1);
+    axes.resize(static_cast<size_t>(id+1));
 
-  assert(axes[id] == 0);
-  axes[id] = std::move(axis);
+  assert(axes[static_cast<size_t>(id)] == 0);
+  axes[static_cast<size_t>(id)] = std::move(axis);
 }
 
 ControllerButton*
 Controller::get_button(int id)
 {
   assert(id >= 0 && id < int(buttons.size()));
-  return buttons[id].get();
+  return buttons[static_cast<size_t>(id)].get();
 }
 
 void
@@ -199,10 +199,10 @@ void
 Controller::add_button(int id, std::unique_ptr<ControllerButton> button)
 {
   if (int(buttons.size())-1 < id)
-    buttons.resize(id+1);
+    buttons.resize(static_cast<size_t>(id+1));
 
-  assert(buttons[id] == 0);
-  buttons[id] = std::move(button);
+  assert(buttons[static_cast<size_t>(id)] == 0);
+  buttons[static_cast<size_t>(id)] = std::move(button);
 }
 
 void
